% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.R
\docType{class}
\name{Run}
\alias{Run}
\title{Run class}
\format{\code{\link{Plate}} class generator}
\usage{
\code{run <- Run$new(xl_file = '~/Google Drive/spectrum/stats/tarnawski/yafa/data/Heat map results YaFa Explant 1.xlsx')}
}
\description{
A class representation of a Bio-Rad plate.
}
\keyword{data}

